# public/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Lician Financial Data API",
    "description": "Free REST API for stock prices, financial statements, insider trades, institutional holdings, and more. Access data for 141,000+ companies across US and European markets.",
    "version": "1.0.0",
    "contact": {
      "name": "Lician API Support",
      "email": "api@lician.com",
      "url": "https://lician.com/developers"
    },
    "license": {
      "name": "API Terms of Service",
      "url": "https://lician.com/terms"
    },
    "x-logo": {
      "url": "https://lician.com/logo.png"
    }
  },
  "servers": [
    {
      "url": "https://lician.com",
      "description": "Production server"
    }
  ],
...
  "paths": {
    "/api/v1/prices/snapshot": {
      "get": {
        "operationId": "getPriceSnapshot",
        "summary": "Get current stock price",
        "description": "Returns real-time price, daily change, volume, and market cap for a stock",
        "tags": ["Prices"],
        "parameters": [
          {
            "name": "ticker",
            "in": "query",
            "required": true,
            "schema": {"type": "string"},
            "description": "Stock ticker symbol (e.g., AAPL, MSFT, GOOGL)",
            "example": "AAPL"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {"$ref": "#/components/schemas/PriceSnapshot"},
                "example": {
                  "snapshot": {
                    "ticker": "AAPL",
                    "price": 185.92,
                    "day_change": 2.34,
                    "day_change_percent": 1.27,
                    "volume": 45234567,
                    "market_cap": 2890000000000
                  }
                }
              }
            }
          },
          "400": {"$ref": "#/components/responses/BadRequest"},
          "404": {"$ref": "#/components/responses/NotFound"},
          "429": {"$ref": "#/components/responses/RateLimited"}
        }
      }
    },
...
    "/api/v1/financials/income-statements": {
      "get": {
        "operationId": "getIncomeStatements",
        "summary": "Get income statements",
        "description": "Returns quarterly or annual income statement data including revenue, gross profit, operating income, and net income",
        "tags": ["Financials"],
        "parameters": [
          {
            "name": "ticker",
            "in": "query",
            "required": true,
            "schema": {"type": "string"},
            "description": "Stock ticker symbol",
            "example": "AAPL"
          },
          {
            "name": "period",
            "in": "query",
            "required": false,
            "schema": {"type": "string", "enum": ["quarterly", "annual"], "default": "quarterly"},
            "description": "Reporting period type"
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {"type": "integer", "default": 4, "maximum": 40},
            "description": "Number of periods to return"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {"$ref": "#/components/schemas/IncomeStatements"},
                "example": {
                  "income_statements": [{
                    "ticker": "AAPL",
                    "report_period": "2024-09-30",
                    "period": "quarterly",
                    "revenue": 94930000000,
                    "cost_of_revenue": 51051000000,
                    "gross_profit": 43879000000,
                    "operating_expenses": 14287000000,
                    "operating_income": 29592000000,
                    "net_income": 14736000000,
                    "eps": 0.97,
                    "eps_diluted": 0.96
                  }]
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/financials/balance-sheets": {
      "get": {
        "operationId": "getBalanceSheets",
        "summary": "Get balance sheets",
        "description": "Returns quarterly or annual balance sheet data including assets, liabilities, and shareholders equity",
        "tags": ["Financials"],
        "parameters": [
          {
            "name": "ticker",
            "in": "query",
            "required": true,
            "schema": {"type": "string"},
            "description": "Stock ticker symbol",
            "example": "AAPL"
          },
          {
            "name": "period",
            "in": "query",
            "required": false,
            "schema": {"type": "string", "enum": ["quarterly", "annual"], "default": "quarterly"},
            "description": "Reporting period type"
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {"type": "integer", "default": 4, "maximum": 40},
            "description": "Number of periods to return"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {"$ref": "#/components/schemas/BalanceSheets"},
                "example": {
                  "balance_sheets": [{
                    "ticker": "AAPL",
                    "report_period": "2024-09-30",
                    "total_assets": 352583000000,
                    "current_assets": 152987000000,
                    "total_liabilities": 290437000000,
                    "current_liabilities": 176392000000,
                    "shareholders_equity": 62146000000,
                    "cash_and_equivalents": 29943000000,
                    "total_debt": 104590000000
                  }]
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/financials/cash-flow-statements": {
      "get": {
        "operationId": "getCashFlowStatements",
        "summary": "Get cash flow statements",
        "description": "Returns quarterly or annual cash flow data including operating, investing, and financing activities",
        "tags": ["Financials"],
        "parameters": [
          {
            "name": "ticker",
            "in": "query",
            "required": true,
            "schema": {"type": "string"},
            "description": "Stock ticker symbol",
            "example": "AAPL"
          },
          {
            "name": "period",
            "in": "query",
            "required": false,
            "schema": {"type": "string", "enum": ["quarterly", "annual"], "default": "quarterly"},
            "description": "Reporting period type"
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {"type": "integer", "default": 4, "maximum": 40},
            "description": "Number of periods to return"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {"$ref": "#/components/schemas/CashFlowStatements"},
                "example": {
                  "cash_flow_statements": [{
                    "ticker": "AAPL",
                    "report_period": "2024-09-30",
                    "operating_cash_flow": 26810000000,
                    "investing_cash_flow": -1923000000,
                    "financing_cash_flow": -29168000000,
                    "free_cash_flow": 23910000000,
                    "capital_expenditures": -2900000000,
                    "dividends_paid": -3800000000
                  }]
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/insider-trades": {
      "get": {
        "operationId": "getInsiderTrades",
        "summary": "Get insider trading activity",
        "description": "Returns SEC Form 4 insider transactions including buys, sells, and option exercises",
        "tags": ["Insider Trading"],
        "parameters": [
          {
            "name": "ticker",
            "in": "query",
            "required": false,
            "schema": {"type": "string"},
            "description": "Filter by stock ticker (optional, returns all if omitted)",
            "example": "AAPL"
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {"type": "integer", "default": 50, "maximum": 500},
            "description": "Number of trades to return"
          },
          {
            "name": "transaction_type",
            "in": "query",
            "required": false,
            "schema": {"type": "string", "enum": ["P-Purchase", "S-Sale", "A-Award", "M-Exercise"]},
            "description": "Filter by transaction type"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {"$ref": "#/components/schemas/InsiderTrades"},
                "example": {
                  "insider_trades": [{
                    "ticker": "AAPL",
                    "insider_name": "Tim Cook",
                    "insider_title": "CEO",
                    "transaction_type": "S-Sale",
                    "shares": 50000,
                    "price": 185.50,
                    "value": 9275000,
                    "shares_owned_after": 3200000,
                    "filing_date": "2024-12-15",
                    "transaction_date": "2024-12-13"
                  }]
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/institutional-ownership": {
      "get": {
        "operationId": "getInstitutionalOwnership",
        "summary": "Get institutional holders",
        "description": "Returns 13F institutional ownership data showing major fund holders",
        "tags": ["Institutional"],
        "parameters": [
          {
            "name": "ticker",
            "in": "query",
            "required": true,
            "schema": {"type": "string"},
            "description": "Stock ticker symbol",
            "example": "AAPL"
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {"type": "integer", "default": 20, "maximum": 100},
            "description": "Number of holders to return"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {"$ref": "#/components/schemas/InstitutionalHoldings"},
                "example": {
                  "holdings": [{
                    "ticker": "AAPL",
                    "holder_name": "Vanguard Group Inc",
                    "shares": 1340000000,
                    "value": 248710000000,
                    "percent_of_shares": 8.72,
                    "change_in_shares": 5200000,
                    "change_percent": 0.39,
                    "report_date": "2024-09-30"
                  }]
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/institutional-ownership/investors": {
      "get": {
        "operationId": "getInstitutionalInvestors",
        "summary": "Get investor portfolio holdings",
        "description": "Returns all holdings for a specific institutional investor from their 13F filing",
        "tags": ["Institutional"],
        "parameters": [
          {
            "name": "investor",
            "in": "query",
            "required": true,
            "schema": {"type": "string"},
            "description": "Investor name or CIK",
            "example": "Berkshire Hathaway"
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {"type": "integer", "default": 50},
            "description": "Number of holdings to return"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {"$ref": "#/components/schemas/InvestorPortfolio"}
              }
            }
          }
        }
      }
    },
    "/api/v1/analyst-ratings": {
      "get": {
        "operationId": "getAnalystRatings",
        "summary": "Get analyst ratings",
        "description": "Returns analyst buy/hold/sell ratings and price targets",
        "tags": ["Analysts"],
        "parameters": [
          {
            "name": "ticker",
            "in": "query",
            "required": true,
            "schema": {"type": "string"},
            "description": "Stock ticker symbol",
            "example": "AAPL"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {"$ref": "#/components/schemas/AnalystRatings"},
                "example": {
                  "ratings": {
                    "ticker": "AAPL",
                    "consensus": "Buy",
                    "buy": 28,
                    "hold": 8,
                    "sell": 2,
                    "average_price_target": 210.50,
                    "high_price_target": 250.00,
                    "low_price_target": 170.00
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/news": {
      "get": {
        "operationId": "getNews",
        "summary": "Get market news",
        "description": "Returns latest financial news headlines and articles",
        "tags": ["News"],
        "parameters": [
          {
            "name": "ticker",
            "in": "query",
            "required": false,
            "schema": {"type": "string"},
            "description": "Filter by stock ticker",
            "example": "AAPL"
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {"type": "integer", "default": 20, "maximum": 100},
            "description": "Number of articles to return"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {"$ref": "#/components/schemas/NewsArticles"}
              }
            }
          }
        }
      }
    },
    "/api/v1/filings": {
      "get": {
        "operationId": "getSECFilings",
        "summary": "Get SEC filings",
        "description": "Returns SEC filings including 10-K, 10-Q, 8-K, and more",
        "tags": ["News"],
        "parameters": [
          {
            "name": "ticker",
            "in": "query",
            "required": true,
            "schema": {"type": "string"},
            "description": "Stock ticker symbol",
            "example": "AAPL"
          },
          {
            "name": "type",
            "in": "query",
            "required": false,
            "schema": {"type": "string", "enum": ["10-K", "10-Q", "8-K", "4", "13F"]},
            "description": "Filter by filing type"
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {"type": "integer", "default": 20},
            "description": "Number of filings to return"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {"$ref": "#/components/schemas/SECFilings"}
              }
            }
          }
        }
      }
    },
...
    "/api/v1/company/facts": {
      "get": {
        "operationId": "getCompanyFacts",
        "summary": "Get company information",
        "description": "Returns company profile including sector, industry, description, and key stats",
        "tags": ["Financials"],
        "parameters": [
          {
            "name": "ticker",
            "in": "query",
            "required": true,
            "schema": {"type": "string"},
            "description": "Stock ticker symbol",
            "example": "AAPL"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {"$ref": "#/components/schemas/CompanyFacts"}
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "PriceSnapshot": {
        "type": "object",
        "properties": {
          "snapshot": {
            "type": "object",
            "properties": {
              "ticker": {"type": "string"},
              "price": {"type": "number"},
              "day_change": {"type": "number"},
              "day_change_percent": {"type": "number"},
              "volume": {"type": "integer"},
              "market_cap": {"type": "integer"},
              "open": {"type": "number"},
              "high": {"type": "number"},
              "low": {"type": "number"},
              "previous_close": {"type": "number"}
            }
          }
        }
      },
...
      "IncomeStatements": {
        "type": "object",
        "properties": {
          "income_statements": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "ticker": {"type": "string"},
                "report_period": {"type": "string", "format": "date"},
                "period": {"type": "string", "enum": ["quarterly", "annual"]},
                "revenue": {"type": "integer"},
                "cost_of_revenue": {"type": "integer"},
                "gross_profit": {"type": "integer"},
                "operating_expenses": {"type": "integer"},
                "operating_income": {"type": "integer"},
                "net_income": {"type": "integer"},
                "eps": {"type": "number"},
                "eps_diluted": {"type": "number"},
                "shares_outstanding": {"type": "integer"}
              }
            }
          }
        }
      },
      "BalanceSheets": {
        "type": "object",
        "properties": {
          "balance_sheets": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "ticker": {"type": "string"},
                "report_period": {"type": "string", "format": "date"},
                "total_assets": {"type": "integer"},
                "current_assets": {"type": "integer"},
                "total_liabilities": {"type": "integer"},
                "current_liabilities": {"type": "integer"},
                "shareholders_equity": {"type": "integer"},
                "cash_and_equivalents": {"type": "integer"},
                "total_debt": {"type": "integer"},
                "retained_earnings": {"type": "integer"}
              }
            }
          }
        }
      },
      "CashFlowStatements": {
        "type": "object",
        "properties": {
          "cash_flow_statements": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "ticker": {"type": "string"},
                "report_period": {"type": "string", "format": "date"},
                "operating_cash_flow": {"type": "integer"},
                "investing_cash_flow": {"type": "integer"},
                "financing_cash_flow": {"type": "integer"},
                "free_cash_flow": {"type": "integer"},
                "capital_expenditures": {"type": "integer"},
                "dividends_paid": {"type": "integer"}
              }
            }
          }
        }
      },
...
      "InsiderTrades": {
        "type": "object",
        "properties": {
          "insider_trades": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "ticker": {"type": "string"},
                "insider_name": {"type": "string"},
                "insider_title": {"type": "string"},
                "transaction_type": {"type": "string"},
                "shares": {"type": "integer"},
                "price": {"type": "number"},
                "value": {"type": "integer"},
                "shares_owned_after": {"type": "integer"},
                "filing_date": {"type": "string", "format": "date"},
                "transaction_date": {"type": "string", "format": "date"}
              }
            }
          }
        }
      },
      "InstitutionalHoldings": {
        "type": "object",
        "properties": {
          "holdings": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "ticker": {"type": "string"},
                "holder_name": {"type": "string"},
                "shares": {"type": "integer"},
                "value": {"type": "integer"},
                "percent_of_shares": {"type": "number"},
                "change_in_shares": {"type": "integer"},
                "change_percent": {"type": "number"},
                "report_date": {"type": "string", "format": "date"}
              }
            }
          }
        }
      },
      "InvestorPortfolio": {
        "type": "object",
        "properties": {
          "investor": {"type": "string"},
          "total_value": {"type": "integer"},
          "holdings": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "ticker": {"type": "string"},
                "company_name": {"type": "string"},
                "shares": {"type": "integer"},
                "value": {"type": "integer"},
                "percent_of_portfolio": {"type": "number"}
              }
            }
          }
        }
      },
      "AnalystRatings": {
        "type": "object",
        "properties": {
          "ratings": {
            "type": "object",
            "properties": {
              "ticker": {"type": "string"},
              "consensus": {"type": "string", "enum": ["Strong Buy", "Buy", "Hold", "Sell", "Strong Sell"]},
              "buy": {"type": "integer"},
              "hold": {"type": "integer"},
              "sell": {"type": "integer"},
              "average_price_target": {"type": "number"},
              "high_price_target": {"type": "number"},
              "low_price_target": {"type": "number"}
            }
          }
        }
      },
...
      "NewsArticles": {
        "type": "object",
        "properties": {
          "articles": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "title": {"type": "string"},
                "summary": {"type": "string"},
                "source": {"type": "string"},
                "url": {"type": "string", "format": "uri"},
                "published_at": {"type": "string", "format": "date-time"},
                "tickers": {"type": "array", "items": {"type": "string"}}
              }
            }
          }
        }
      },
      "SECFilings": {
        "type": "object",
        "properties": {
          "filings": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "ticker": {"type": "string"},
                "form_type": {"type": "string"},
                "filing_date": {"type": "string", "format": "date"},
                "description": {"type": "string"},
                "url": {"type": "string", "format": "uri"}
              }
            }
          }
        }
      },
...
    },
    "responses": {
      "BadRequest": {
        "description": "Bad request - missing or invalid parameters",
        "content": {
          "application/json": {
            "schema": {"$ref": "#/components/schemas/Error"},
            "example": {"error": "Bad Request", "message": "Missing required parameter: ticker", "code": 400}
          }
        }
      },
      "NotFound": {
        "description": "Resource not found",
        "content": {
          "application/json": {
            "schema": {"$ref": "#/components/schemas/Error"},
            "example": {"error": "Not Found", "message": "No data found for ticker: INVALID", "code": 404}
          }
        }
      },
      "RateLimited": {
        "description": "Rate limit exceeded",
        "content": {
          "application/json": {
            "schema": {"$ref": "#/components/schemas/Error"},
            "example": {"error": "Too Many Requests", "message": "Rate limit exceeded. Limit: 100 requests/minute", "code": 429}
          }
        }
      }
    },
    "securitySchemes": {
      "ApiKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key",
        "description": "API key for authenticated access. Get one at https://lician.com/developers/signup"
      }
    }
  },
  "security": [],
...
}